% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{callAPI}
\alias{callAPI}
\title{Call API}
\usage{
callAPI(
  action = c("current_package_list_with_resources", "group_list", "package_list",
    "organization_list", "tag_list"),
  ...
)
}
\arguments{
\item{action}{(character) name of the endpoint
"mapping"}

\item{...}{parameters for the endpoint, e.g. all_fields = "true"}
}
\value{
(data.frame) output from the Pandora API
}
\description{
Call API
}
