% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{filterPattern}
\alias{filterPattern}
\title{Filter Pattern}
\usage{
filterPattern(datAPI, pattern = "")
}
\arguments{
\item{datAPI}{(list) output from the Pandora API}

\item{pattern}{(character) string for filtering all meta information}
}
\value{
(list) a data.frame with rows that contain the pattern
}
\description{
Search for pattern in all columns of datAPI and filter respective rows
}
