% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{filterResourceByName}
\alias{filterResourceByName}
\title{Filter Resource by Name}
\usage{
filterResourceByName(resource, name)
}
\arguments{
\item{resource}{(data.frame) resources data frame}

\item{name}{(character) name of a resource}
}
\value{
(data.frame) filtered resource
}
\description{
Filter Resource by Name
}
