% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{filterValidFileType}
\alias{filterValidFileType}
\title{Filter Resource by Valid File Type}
\usage{
filterValidFileType(resource, name)
}
\arguments{
\item{resource}{(data.frame) resources data frame}

\item{name}{(character) name of a resource}
}
\value{
(data.frame) filtered resource
}
\description{
Filter Resource by Valid File Type
}
