% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{formatRepositoryList}
\alias{formatRepositoryList}
\title{Rename Repository Meta Columns}
\usage{
formatRepositoryList(
  packageList,
  columns = getDatasetFields(),
  renameColumns = TRUE
)
}
\arguments{
\item{packageList}{(data.frame) optional, output of callAPI() e.g. from a previous call to the
Pandora API.}

\item{columns}{(character) names of columns that should be returned}

\item{renameColumns}{(logical) apply names from the 'Additional Info' box from
'https://pandoradata.earth/dataset/' to the columns of returned data}
}
\value{
(data.frame) containing available repositories
}
\description{
Apply names from the 'Additional Info' box from 'https://pandoradata.earth/dataset/' to the
columns of returned data
}
