% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getDatasetFields}
\alias{getDatasetFields}
\title{Get Dataset Meta Fields}
\usage{
getDatasetFields()
}
\value{
(character vector) names of meta fields
}
\description{
Names of particular meta fields from the 'Additional Info' box from 'https://pandoradata.earth/dataset/'
}
