% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getNetworks}
\alias{getNetworks}
\title{Get Networks}
\usage{
getNetworks(pattern = "", order = TRUE, groupList = data.frame())
}
\arguments{
\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE, order dataframe alphabetically by 'repository' and 'name'}

\item{groupList}{(data.frame) optional, output of callAPI() from a previous call to the
Pandora API.}
}
\value{
(data.frame) giving the "name" and "display_name" of available Pandora
networks (groups in CKAN terminology)
}
\description{
Get all available networks (groups in CKAN terminology)
optional filtering of names for a given string
}
