% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{getNrow}
\alias{getNrow}
\title{get nRow}
\usage{
getNrow(type, nrows = NA_integer_)
}
\arguments{
\item{type}{(character) file type}

\item{nrows}{integer: the maximum number of rows to read in.  Negative
    and other invalid values are ignored.}
}
\description{
get nRow
}
