% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getRepositories}
\alias{getRepositories}
\title{Get Repositories}
\usage{
getRepositories(
  network = "",
  pattern = "",
  order = TRUE,
  columns = getDatasetFields(),
  renameColumns = TRUE,
  packageList = data.frame()
)
}
\arguments{
\item{network}{(character) name of a Pandora network, e.g. an entry of the output from
\code{getNetworks()$name}}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE, order dataframe alphabetically by 'repository' and 'name'}

\item{columns}{(character) names of columns that should be returned}

\item{renameColumns}{(logical) apply names from the 'Additional Info' box from
'https://pandoradata.earth/dataset/' to the columns of returned data}

\item{packageList}{(data.frame) optional, output of callAPI() e.g. from a previous call to the
Pandora API.}
}
\value{
(data.frame) containing available repositories
}
\description{
Get all vailable repositories or those within
a specific network
optional filtering of meta information for a given string
}
