% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-ckan.R
\name{getResources}
\alias{getResources}
\title{Get Resources}
\usage{
getResources(
  fileType = character(),
  repository = "",
  network = "",
  pattern = "",
  order = TRUE,
  packageList = data.frame()
)
}
\arguments{
\item{fileType}{(character) list of relevant file types, e.g. c("xls", "xlsx", "csv", "odt")}

\item{repository}{(character) name of a Pandora repository, e.g. an entry of the output from
\code{getRepositories()$name}}

\item{network}{(character) name of a Pandora network, e.g. an entry of the output from
\code{getNetworks()$name}}

\item{pattern}{(character) string for meta information search}

\item{order}{(logical) if TRUE, order dataframe alphabetically by 'repository' and 'name'}

\item{packageList}{(data.frame) optional, output of callAPI() e.g. from a previous call to the
Pandora API.}
}
\value{
(data.frame) containing available resources within a repository
}
\description{
Get all available resources within a repository or filtered
by file type or those within a specific network or within a specific repository
optional filtering of meta information for a given string
}
