% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_genes.R
\name{pair_genes}
\alias{pair_genes}
\title{Pair Genes for Subtype Prediction}
\usage{
pair_genes(expression_matrix)
}
\arguments{
\item{expression_matrix}{A matrix where rows are genes and columns are samples.}

\item{goal_gene_pairs}{A vector of gene pairs in the format "geneA.geneB".}
}
\value{
A dataframe of paired genes with values 0 or 1, where 1 means geneA > geneB.
}
\description{
This function pairs genes from the expression matrix based on the goal gene pairs.
}
