\name{Pandora-package}
\alias{Pandora-package}
\alias{Pandora}
\docType{package}
\title{
Data processing of cut marks in bones, from images of sections.
}
\description{
According "Gonzalez, M. A. M., Yravedra, J., Gonzalez-Aguilera, D., Palomeque-GonzAlez, J. F., & Dominguez-Rodrigo, M. (2015). Micro-photogrammetric characterization of cut marks on bones. Journal of Archaeological Science, 62, 128-142.", this package facilites processing of data from images of sections (in .JPG format) of cut marks in bones, from 3D virtual models.
}
\details{
\tabular{ll}{
Package: \tab Pandora\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-15\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
Juan F. Palomeque-gonzalez<jfpalomeque.gonzalez@gmail.com>

}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
