% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_molecular_subtype.R
\name{predict_molecular_subtype}
\alias{predict_molecular_subtype}
\title{Predict Molecular Subtypes}
\usage{
predict_molecular_subtype(expression_matrix, type = "RNA-seq")
}
\arguments{
\item{expression_matrix}{A matrix where rows are genes and columns are samples.}

\item{type}{A string specifying the type of input data: "RNA-seq" (default) or "PCR".}
}
\value{
A dataframe with two columns: \code{SampleName} and \code{PredictedSubtype}.
}
\description{
This function predicts molecular subtypes based on the input expression matrix.
It supports both RNA-seq and PCR data types.
}
