% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_subtype.R
\name{predict_subtype}
\alias{predict_subtype}
\title{Predict Molecular Subtypes Using a Pre-trained Model}
\usage{
predict_subtype(paired_genesT, model = NULL)
}
\arguments{
\item{paired_genesT}{A transposed dataframe of paired genes with values 0 or 1.}

\item{model}{A pre-trained machine learning model. If NULL, the model will be loaded from the package.}
}
\value{
A vector of predicted subtypes.
}
\description{
This function uses a pre-trained model to predict molecular subtypes based on the paired gene data.
}
