% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{selectSingleFile}
\alias{selectSingleFile}
\title{Select Single File from Resources}
\usage{
selectSingleFile(resource)
}
\arguments{
\item{resource}{(data.frame) resources data frame}
}
\value{
(data.frame) selected resource
}
\description{
Select Single File from Resources
}
