% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-getData.R
\name{validateResource}
\alias{validateResource}
\title{Validate Resource}
\usage{
validateResource(resource, repository)
}
\arguments{
\item{resource}{(data.frame) resources data frame}

\item{repository}{(character) name of a Pandora repository, e.g. an entry of the output from
\code{getRepositories()$name}}
}
\value{
(data.frame) resource, or error if empty
}
\description{
Validate Resource
}
